+++ frontend_qt/mainwindow.h
@@ -136,19 +136,22 @@ protected:
 	void combobox_item_enabled(QComboBox *comboBox, int index, bool enabled);
     void upcean_addon_gap(QComboBox *comboBox, QLabel *label, int base);
     void set_gs1_mode(bool gs1_mode);
+
 	const char *get_setting_name(int symbology);
+
 	int get_button_group_index(const QStringList &children);
 	void set_radiobutton_from_setting(QSettings &settings, const QString &setting, const QStringList &children, int default_val = 0);
 	int get_combobox_index(const QString &child);
 	void set_combobox_from_setting(QSettings &settings, const QString &setting, const QString &child, int default_val = 0);
 	int get_checkbox_val(const QString &child);
 	void set_checkbox_from_setting(QSettings &settings, const QString &setting, const QString &child, int default_val = 0);
-	float get_doublespinbox_val(const QString &child);
-	void set_doublespinbox_from_setting(QSettings &settings, const QString &setting, const QString &child, float default_val = 0);
+	double get_doublespinbox_val(const QString &child);
+	void set_doublespinbox_from_setting(QSettings &settings, const QString &setting, const QString &child, float default_val = 0.0f);
 	QString get_lineedit_val(const QString &child);
 	void set_lineedit_from_setting(QSettings &settings, const QString &setting, const QString &child, const char *default_val = "");
 	int get_spinbox_val(const QString &child);
 	void set_spinbox_from_setting(QSettings &settings, const QString &setting, const QString &child, int default_val = 0);
+
 	void save_sub_settings(QSettings &settings, int symbology);
 	void load_sub_settings(QSettings &settings, int symbology);
